+++ crates/zng/src/lib.rs
@@ -598,7 +598,7 @@ mod __prelude {
 
     pub use zng_clone_move::{async_clmv, async_clmv_fn, async_clmv_fn_once, clmv};
 
-    pub use zng_color::{colors, hex, hsl, hsla, hsv, hsva, rgb, rgba, web_colors, MixAdjust as _};
+    pub use zng_color::{colors, hex, hsl, hsla, hsv, hsva, rgb, rgba, rgba_pair, web_colors, MixAdjust as _, RgbaPairVarExt as _};
 
     pub use zng_ext_clipboard::CLIPBOARD;
 
@@ -781,8 +781,8 @@ mod __prelude_wgt {
     pub use zng_wgt::{wgt_fn, WidgetFn};
 
     pub use zng_color::{
-        color_scheme_highlight, color_scheme_map, color_scheme_pair, colors, gradient, hex, hsl, hsla, hsv, hsva, rgb, rgba, web_colors,
-        ColorPair, ColorScheme, Hsla, Hsva, MixAdjust as _, MixBlendMode, Rgba,
+        colors, gradient, hex, hsl, hsla, hsv, hsva, rgb, rgba, rgba_pair, web_colors, ColorScheme, Hsla, Hsva, MixAdjust as _,
+        MixBlendMode, Rgba, RgbaPair,
     };
 
     pub use zng_wgt::node::{