+++ examples/icon/src/main.rs
@@ -5,10 +5,7 @@
 use zng::{
     access::ACCESS,
     clipboard,
-    color::{
-        color_scheme_map,
-        filter::{backdrop_blur, drop_shadow, opacity},
-    },
+    color::filter::{backdrop_blur, drop_shadow, opacity},
     container,
     data_view::{DataView, DataViewArgs},
     focus::{directional_nav, focus_scope, focus_shortcut, tab_nav, DirectionalNav, TabNav},
@@ -201,7 +198,7 @@ fn expanded_icon(name: &'static str, ico: icon::GlyphIcon, font_mod: &'static st
         id = "expanded-icon";
         widget::modal = true;
         backdrop_blur = 2;
-        background_color = color_scheme_map(colors::WHITE.with_alpha(10.pct()), colors::BLACK.with_alpha(10.pct()));
+        background_color = rgba_pair(colors::WHITE.with_alpha(10.pct()), colors::BLACK.with_alpha(10.pct()));
         child_align = Align::CENTER;
         on_click = hn!(|args: &ClickArgs| {
             if WIDGET.id() == args.target.widget_id() {
@@ -211,7 +208,7 @@ fn expanded_icon(name: &'static str, ico: icon::GlyphIcon, font_mod: &'static st
         });
         child = Container! {
             id = "panel";
-            background_color = color_scheme_map(colors::BLACK.with_alpha(90.pct()), colors::WHITE.with_alpha(90.pct()));
+            background_color = rgba_pair(colors::BLACK.with_alpha(90.pct()), colors::WHITE.with_alpha(90.pct()));
             focus_scope = true;
             tab_nav = TabNav::Cycle;
             directional_nav = DirectionalNav::Cycle;
@@ -240,7 +237,7 @@ fn expanded_icon(name: &'static str, ico: icon::GlyphIcon, font_mod: &'static st
                                             ico = ico.clone();
                                             ico_size = size;
 
-                                            background_color = color_scheme_map(
+                                            background_color = rgba_pair(
                                                 colors::BLACK.with_alpha(85.pct()),
                                                 colors::WHITE.with_alpha(85.pct())
                                             );