+++ examples/layer/src/main.rs
@@ -2,7 +2,7 @@
 
 use zng::{
     button,
-    color::{color_scheme_map, filter::opacity},
+    color::filter::opacity,
     focus::{directional_nav, focus_scope, tab_nav, DirectionalNav, TabNav},
     layer::AnchorOffset,
     layout::{align, margin, offset},
@@ -62,14 +62,14 @@ fn overlay(id: impl Into<WidgetId>, offset: i32) -> impl UiNode {
     Container! {
         id;
         widget::modal = true;
-        background_color = color_scheme_map(colors::WHITE.with_alpha(10.pct()), colors::BLACK.with_alpha(10.pct()));
+        background_color = rgba_pair(colors::WHITE.with_alpha(10.pct()), colors::BLACK.with_alpha(10.pct()));
         child_align = Align::CENTER;
         child = Container! {
             offset = (offset, offset);
             focus_scope = true;
             tab_nav = TabNav::Cycle;
             directional_nav = DirectionalNav::Cycle;
-            background_color = color_scheme_map(
+            background_color = rgba_pair(
                 colors::GREEN.darken(80.pct()),
                 colors::WHITE.with_alpha(80.pct()).mix_normal(colors::GREEN)
             );